Add Reminders and adjust UI
Reminders have not been added to the model and logic of ePoch. The UI is
also using the old design and has not yet been updated to factor in the
3 columns needed.

Let's add Reminders to the overall logic of ePoch so that the UI can be
properly formatted and tested. Let's also update the UI to include the
columns needed to display Persons, CCAs, and Reminders.